// 4-factory-method-pattern/user/types/enums/permissions.enum.ts
export enum Permissions {
  READ_USER = 'READ_USER',
  CREATE_USER = 'CREATE_USER',
  UPDATE_USER = 'UPDATE_USER',
  DELETE_USER = 'DELETE_USER',
  READ_PROJECT = 'READ_PROJECT',
  CREATE_PROJECT = 'CREATE_PROJECT',
  UPDATE_PROJECT = 'UPDATE_PROJECT',
  DELETE_PROJECT = 'DELETE_PROJECT',
  READ_PROJECT_TASK = 'READ_PROJECT_TASK',
  CREATE_PROJECT_TASK = 'CREATE_PROJECT_TASK',
  UPDATE_PROJECT_TASK = 'UPDATE_PROJECT_TASK',
  DELETE_PROJECT_TASK = 'DELETE_PROJECT_TASK',
  ASSIGN_PROJECT_TASK = 'ASSIGN_PROJECT_TASK',
  UNASSIGN_PROJECT_TASK = 'UNASSIGN_PROJECT_TASK',
  READ_PROJECT_TASK_COMMENT = 'READ_PROJECT_TASK_COMMENT',
  CREATE_PROJECT_TASK_COMMENT = 'CREATE_PROJECT_TASK_COMMENT',
  UPDATE_PROJECT_TASK_COMMENT = 'UPDATE_PROJECT_TASK_COMMENT',
  DELETE_PROJECT_TASK_COMMENT = 'DELETE_PROJECT_TASK_COMMENT',
}